test - renaming the file/url/page name to remove the word (static)
