+++ src/Views/Components/Menu/index.js
@@ -80,9 +80,9 @@ export default function Menu(props) {
             <Link to="/about/">About me</Link>
           </li>
         )}
-        {!window.location.href.includes('static') && (
+        {!window.location.href.includes('arts') && (
           <li>
-            <Link to="/static/">Static arts</Link>
+            <Link to="/arts/">Static Arts</Link>
           </li>
         )}
         {!window.location.href.includes('animations') && (
